requirements.txt
Flask[async]==2.2.2
keras==2.10.0
numpy==1.23.4
Pillow==9.3.0
waitress==2.1.2
tensorflow==2.10.0
